% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poverty-package.R
\docType{package}
\name{poverty-package}
\alias{poverty}
\alias{poverty-package}
\title{poverty: Analyse de la pauvrete}
\description{
Analyse de la pauvrete.
}
\author{
\strong{Maintainer}: Prosper LAWA \email{lawafoum@gmail.com}

Authors:
\itemize{
  \item Samba DIENG \email{sambadiengys122003@gmail.com}
}

}
\keyword{internal}
